+++ CHANGELOG.md
@@ -6,6 +6,7 @@ It is based on Tidepool Blip 1.27.
 ### Changed
 - YLP-3478 Update the front of the maintenance page
 - YLP-3522 Handle accounts marked for deletion
+- YLP-3547 Swap unit and percentage in insulin card
 
 ### Fixed
 - YLP-3290 Top banner is hiding the top part of note creation popup